// src/app/models/Species.ts
export class Species {
    name: string;
    classification: string;
    designation: string;
    // tslint:disable-next-line:variable-name
    average_height: string;
    // tslint:disable-next-line:variable-name
    skin_colors: string;
    // tslint:disable-next-line:variable-name
    hair_colors: string;
    // tslint:disable-next-line:variable-name
    eye_colors: string;
    // tslint:disable-next-line:variable-name
    average_lifespan: string;
    homeworld: string;
    language: string;
    people: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}