// src/app/models/Starship.ts
export class Starship {
    name: string;
    model: string;
    // tslint:disable-next-line:variable-name
    cost_in_credits: string;
    length: string;
    // tslint:disable-next-line:variable-name
    max_atmosphering_speed: string;
    crew: string;
    passengers: string;
    // tslint:disable-next-line:variable-name
    cargo_capacity: string;
    consumables: string;
    // tslint:disable-next-line:variable-name
    hyperdrive_rating: string;
    MGLT: string;
    // tslint:disable-next-line:variable-name
    starship_class: string;
    pilots: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}