// src/app/models/Vehicle.ts
export class Vehicle {
    name: string;
    model: string;
    manufacturer: string;
    // tslint:disable-next-line:variable-name
    cost_in_credits: string;
    length: string;
    // tslint:disable-next-line:variable-name
    max_atmosphering_speed: string;
    crew: string;
    passengers: string;
    // tslint:disable-next-line:variable-name
    cargo_capacity: string;
    consumables: string;
    // tslint:disable-next-line:variable-name
    vehicle_class: string;
    pilots: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}