// src/app/models/Planet.ts
export class Planet {
    name: string;
    // tslint:disable-next-line:variable-name
    rotation_period: string;
    // tslint:disable-next-line:variable-name
    orbital_period: string;
    diameter: string;
    climate: string;
    gravity: string;
    terrain: string;
    // tslint:disable-next-line:variable-name
    surface_water: string;
    population: string;
    residents: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}